renamed class name, added unit test boilerplate
